# api/openapi.yaml
openapi: 3.1.0
info:
  title: Media Hub API
  version: 0.16.0
  description: Control-plane API for private media search, subscriptions, and automation.
servers:
  - url: /api/v1
paths:
...
  /auth/login:
    post:
      operationId: login
      summary: Create a revocable Web or Android session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated session. Web sessions are also set as HttpOnly cookies.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/ProblemResponse'
        '401':
          $ref: '#/components/responses/ProblemResponse'
        '429':
          $ref: '#/components/responses/ProblemResponse'
  /auth/session:
    get:
      operationId: getSession
      summary: Read the current authenticated session
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Current session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '401':
          $ref: '#/components/responses/ProblemResponse'
  /auth/logout:
    post:
      operationId: logout
      summary: Revoke the current session
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '204':
          description: Session revoked
        '401':
          $ref: '#/components/responses/ProblemResponse'
        '403':
          $ref: '#/components/responses/ProblemResponse'
  /auth/password:
    put:
      operationId: changePassword
      summary: Change the administrator password and revoke other sessions
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangePasswordRequest' }
      responses:
        '204': { description: Password changed }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '403': { $ref: '#/components/responses/ProblemResponse' }
        '422': { $ref: '#/components/responses/ProblemResponse' }
  /client/android/releases/latest:
    get:
      operationId: getLatestAndroidRelease
      summary: Read the latest private Android release metadata
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: Latest signed Android release
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AndroidRelease' }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '404': { $ref: '#/components/responses/ProblemResponse' }
        '503': { $ref: '#/components/responses/ProblemResponse' }
  /client/android/releases/{versionCode}/apk:
    get:
      operationId: downloadAndroidRelease
      summary: Download one authenticated signed Android APK
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - name: versionCode
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Signed Android APK
          headers:
            X-Checksum-SHA256:
              schema: { type: string, pattern: '^[0-9a-f]{64}$' }
          content:
            application/vnd.android.package-archive:
              schema: { type: string, format: binary }
        '400': { $ref: '#/components/responses/ProblemResponse' }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '404': { $ref: '#/components/responses/ProblemResponse' }
        '503': { $ref: '#/components/responses/ProblemResponse' }
  /system/overview:
    get:
      operationId: getSystemOverview
      summary: Read current integration health
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Integration overview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemOverview'
        '401':
          $ref: '#/components/responses/ProblemResponse'
  /settings/providers:
    get:
      operationId: getProviderSettings
      summary: Read non-secret runtime provider settings and secret-presence flags
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: Runtime provider settings without secret values
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProviderSettings' }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '503': { $ref: '#/components/responses/ProblemResponse' }
    put:
      operationId: updateProviderSettings
      summary: Atomically encrypt, persist, and apply runtime provider settings
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/CSRFToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProviderSettingsUpdate' }
      responses:
        '200':
          description: Updated settings without secret values
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProviderSettings' }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '403': { $ref: '#/components/responses/ProblemResponse' }
        '409': { $ref: '#/components/responses/ProblemResponse' }
        '422': { $ref: '#/components/responses/ProblemResponse' }
        '503': { $ref: '#/components/responses/ProblemResponse' }
  /statistics/summary:
    get:
      operationId: getStatisticsSummary
      summary: Read user-scoped Media Hub operational counters
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: Operational counters without provider-private data
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperationalStatistics' }
        '401': { $ref: '#/components/responses/ProblemResponse' }
  /discovery/trending:
    get:
      operationId: getTrending
      summary: Read normalized weekly TMDB trending identities
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - name: mediaType
          in: query
          schema: { type: string, enum: [all, movie, series], default: all }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 20, default: 20 }
      responses:
        '200':
          description: Trending media identities
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DiscoveryResponse' }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '503': { $ref: '#/components/responses/ProblemResponse' }
  /discovery/{mediaType}/{tmdbId}/recommendations:
    get:
      operationId: getRecommendations
      summary: Read normalized TMDB recommendations for one identity
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - name: mediaType
          in: path
          required: true
          schema: { type: string, enum: [movie, series] }
        - name: tmdbId
          in: path
          required: true
          schema: { type: string, pattern: '^[1-9][0-9]*$' }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 20, default: 20 }
      responses:
        '200':
          description: Recommended media identities
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DiscoveryResponse' }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '503': { $ref: '#/components/responses/ProblemResponse' }
  /search:
    get:
      operationId: searchMedia
      summary: Search normalized resource candidates
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 120
      responses:
        '200':
          description: Search response, including per-source partial failures
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/ProblemResponse'
        '401':
          $ref: '#/components/responses/ProblemResponse'
  /integrations/wecom/test:
    post:
      operationId: testWeComNotification
      summary: Send one fixed test notification through the current Enterprise WeChat configuration
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/CSRFToken' }
      responses:
        '200':
          description: Test notification accepted by Enterprise WeChat
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties: { status: { type: string, enum: [sent] } }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '502': { $ref: '#/components/responses/ProblemResponse' }
        '503': { $ref: '#/components/responses/ProblemResponse' }
  /integrations/qmediasync/status:
    get:
      operationId: getQMediaSyncStatus
      summary: Read QMediaSync version and recent synchronization records
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: QMediaSync status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QMediaSyncStatus'
        '401':
          $ref: '#/components/responses/ProblemResponse'
        '502':
          $ref: '#/components/responses/ProblemResponse'
        '503':
          $ref: '#/components/responses/ProblemResponse'
        '504':
          $ref: '#/components/responses/ProblemResponse'
  /integrations/emby/libraries:
    get:
      operationId: getEmbyLibraries
      summary: Read configured Emby media libraries
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Emby libraries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmbyLibraryResponse'
        '401':
          $ref: '#/components/responses/ProblemResponse'
        '502':
          $ref: '#/components/responses/ProblemResponse'
        '503':
          $ref: '#/components/responses/ProblemResponse'
        '504':
          $ref: '#/components/responses/ProblemResponse'
  /integrations/emby/items:
    get:
      operationId: searchEmbyItems
      summary: Search Emby for duplicate movies or series
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 120
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Matching Emby items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmbyItemSearchResponse'
        '400':
          $ref: '#/components/responses/ProblemResponse'
        '401':
          $ref: '#/components/responses/ProblemResponse'
        '502':
          $ref: '#/components/responses/ProblemResponse'
        '503':
          $ref: '#/components/responses/ProblemResponse'
        '504':
          $ref: '#/components/responses/ProblemResponse'
  /integrations/emby/libraries/{id}/items:
    get:
      operationId: browseEmbyLibraryItems
      summary: Browse movies and series in one Emby library
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, pattern: '^[A-Za-z0-9_-]{1,128}$' } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, maximum: 1000000, default: 0 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
      responses:
        '200':
          description: Emby library items
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmbyItemSearchResponse' }
        '400': { $ref: '#/components/responses/ProblemResponse' }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '502': { $ref: '#/components/responses/ProblemResponse' }
        '503': { $ref: '#/components/responses/ProblemResponse' }
        '504': { $ref: '#/components/responses/ProblemResponse' }
  /integrations/emby/libraries/{id}/refresh:
    post:
      operationId: refreshEmbyLibrary
      summary: Ask Emby to refresh one media library
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, pattern: '^[A-Za-z0-9_-]{1,128}$' } }
        - { $ref: '#/components/parameters/CSRFToken' }
      responses:
        '202':
          description: Emby refresh accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmbyRefreshResponse' }
        '400': { $ref: '#/components/responses/ProblemResponse' }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '403': { $ref: '#/components/responses/ProblemResponse' }
        '502': { $ref: '#/components/responses/ProblemResponse' }
        '503': { $ref: '#/components/responses/ProblemResponse' }
        '504': { $ref: '#/components/responses/ProblemResponse' }
  /integrations/emby/items/{id}:
    get:
      operationId: getEmbyItem
      summary: Read safe metadata for one Emby item
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, pattern: '^[A-Za-z0-9_-]{1,128}$' } }
      responses:
        '200':
          description: Emby item metadata without file paths or credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmbyItemDetail' }
        '400': { $ref: '#/components/responses/ProblemResponse' }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '404': { $ref: '#/components/responses/ProblemResponse' }
        '502': { $ref: '#/components/responses/ProblemResponse' }
        '503': { $ref: '#/components/responses/ProblemResponse' }
        '504': { $ref: '#/components/responses/ProblemResponse' }
  /integrations/emby/items/{id}/refresh:
    post:
      operationId: refreshEmbyItem
      summary: Ask Emby to refresh one media item
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, pattern: '^[A-Za-z0-9_-]{1,128}$' } }
        - { $ref: '#/components/parameters/CSRFToken' }
      responses:
        '202':
          description: Emby refresh accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmbyRefreshResponse' }
        '400': { $ref: '#/components/responses/ProblemResponse' }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '403': { $ref: '#/components/responses/ProblemResponse' }
        '502': { $ref: '#/components/responses/ProblemResponse' }
        '503': { $ref: '#/components/responses/ProblemResponse' }
        '504': { $ref: '#/components/responses/ProblemResponse' }
  /integrations/115/status:
    get:
      operationId: getDrive115Status
      summary: Check 115 QR authorization without returning account identity
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: 115 authorization and storage status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Drive115Status'
        '401':
          $ref: '#/components/responses/ProblemResponse'
        '502':
          $ref: '#/components/responses/ProblemResponse'
        '503':
          $ref: '#/components/responses/ProblemResponse'
        '504':
          $ref: '#/components/responses/ProblemResponse'
  /integrations/115/auth/device:
    post:
      operationId: startDrive115DeviceAuthorization
      summary: Start encrypted 115 QR login without an Open Platform client id
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/CSRFToken' }]
      responses:
        '201':
          description: Device authorization with an authenticated-only QR image
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Drive115DeviceAuthorization' }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '403': { $ref: '#/components/responses/ProblemResponse' }
        '503': { $ref: '#/components/responses/ProblemResponse' }
  /integrations/115/auth/device/{id}:
    get:
      operationId: pollDrive115DeviceAuthorization
      summary: Poll one user-scoped 115 QR authorization
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, minLength: 16, maxLength: 128 }
      responses:
        '200':
          description: Current device authorization state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Drive115DeviceAuthorization' }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '404': { $ref: '#/components/responses/ProblemResponse' }
  /integrations/115/files:
    get:
      operationId: listDrive115Files
      summary: List one 115 directory without exposing download URLs or account identity
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: parentId, in: query, schema: { type: string, pattern: '^[0-9]+$', default: '0' } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
      responses:
        '200':
          description: Bounded file listing
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Drive115FileList' }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '502': { $ref: '#/components/responses/ProblemResponse' }
  /playback/descriptors:
    post:
      operationId: createPlaybackDescriptor
      summary: Resolve one trusted 115 video file into an upstream temporary playback description
      description: The upstream URL lifetime may be unknown. Clients must re-resolve the same parentId/fileId after a recognized expiration response.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/CSRFToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlaybackDescriptorRequest' }
      responses:
        '201':
          description: Upstream-issued playback description; Media Hub does not proxy media bytes
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlaybackDescriptor' }
        '400': { $ref: '#/components/responses/ProblemResponse' }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '404': { $ref: '#/components/responses/ProblemResponse' }
        '502': { $ref: '#/components/responses/ProblemResponse' }
        '503': { $ref: '#/components/responses/ProblemResponse' }
...
  /integrations/115/commands/{id}:
    get:
      operationId: getDrive115Command
      summary: Read one user-scoped 115 command and its event history
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          description: Command detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Drive115Command' }
        '404': { $ref: '#/components/responses/ProblemResponse' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: media_hub_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: false
      description: Required for state-changing requests authenticated by Web cookie.
      schema:
        type: string
...
  responses:
    ProblemResponse:
      description: Request failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [password, client]
      properties:
        password:
          type: string
          format: password
          minLength: 1
          maxLength: 1024
        client:
          type: string
          enum: [web, android]
    LoginResponse:
      type: object
      required: [authenticated, client, expiresAt]
      properties:
        authenticated:
          type: boolean
          const: true
        client:
          type: string
          enum: [web, android]
        expiresAt:
          type: string
          format: date-time
        token:
          type: string
          description: Returned only for Android sessions.
        csrfToken:
          type: string
          description: Returned only for Web sessions.
    SessionResponse:
      type: object
      required: [authenticated, client, expiresAt]
      properties:
        authenticated:
          type: boolean
          const: true
        client:
          type: string
          enum: [web, android]
        expiresAt:
          type: string
          format: date-time
...